Move non-nodes to seperate package
