+++ src/main/java/org/triplea/client/ui/javafx/MainMenuPane.java
@@ -4,6 +4,8 @@ import java.io.IOException;
 import java.text.MessageFormat;
 import java.util.function.Function;
 
+import org.triplea.client.ui.javafx.util.FxmlManager;
+
 import games.strategy.engine.ClientContext;
 import games.strategy.triplea.UrlConstants;
 import javafx.beans.binding.Bindings;
